create profile update

+++ app/Http/Controllers/UserController.php
@@ -38,9 +38,9 @@ class UserController extends Controller
         $user = User::where('username', $request->input('username'))->get()->first();
 
         if ($user) return response()->json(['message' => 'Admin dengan username telah tersedia!'], 500);
-        
+
         try {
-            
+
             User::create([
                 'username' => $request->input('username'),
                 'name' => $request->input('name'),
@@ -56,18 +56,48 @@ class UserController extends Controller
 
     public function updateProfile(Request $request)
     {
-        $user = $this->getUserLogin();
-        $path = $request->file('image')->store('user_profile');
+        $request->validate([
+            "image" => "nullable",
+            "username" => "nullable",
+            "nip" => "nullable",
+            "name" => "nullable",
+            "old_password" => "nullable",
+            "new_password" => "nullable",
+        ]);
 
+        $user = $this->getUserLogin();
+
+        // dd($request->file("image")->storeAs("/user_profile", $fileName));
         if ($user->role) {
             $getUser = User::where('id', $user->id)->first();
-            // dd($getUser->id);
-            Storage::delete($getUser->image);
             $getUser->update([
                 'username' => $request->input('username') ? $request->input('username') : $getUser->username,
                 'name' => $request->input('name') ? $request->input('name') : $getUser->name,
-                'image' => $request->file('image') ? $path : $getUser->image,
             ]);
+
+            // if user change image
+            if ($request->file("image")) {
+                $file = $request->file("image");
+                $fileName = time() . "-" . $user->username . "." . $file->getClientOriginalExtension();
+
+                $path = $request->file("image")->storeAs("/user_profile", $fileName);
+                if ($getUser->image) Storage::delete($getUser->image);
+
+                $getUser->update([
+                    "image" => $path,
+                ]);
+            };
+
+            // if user change password
+            if ($request->input("new_password")) {
+                if (!Hash::check($request->input("old_password"), $getUser->password)) return response()->json(["message" => "Old Password tidak sama!"], 500);
+
+                $getUser->update([
+                    "password" => Hash::make($request->input("new_password")),
+                ]);
+            }
+
+            return response()->json(["message" => "Data berhasil di update"], 200);
         }
 
         // $getUser = Member::where('id_member', $user->id_member)->first();
